server/requirements.txt
torch
transformers
pyannote.audio
spacy
yake
pymongo
google-cloud-storage
firebase-admin
google-generativeai
pydub
scipy